+++ src/z3utils.rs
